Some refactorings for easier viewing of components
